requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
python-dotenv==1.0.0
python-multipart==0.0.9
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-vision==3.7.2
googleapis-common-protos==1.63.0
google-api-python-client==2.126.0
google-generativeai==0.5.2
google-cloud-aiplatform==1.48.0